requirements.txt
# Core Google ADK - Agent Development Kit
google-adk

# Web and API frameworks
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data processing and storage
pandas>=2.1.0
numpy>=1.24.0
sqlalchemy>=2.0.0

# Vector databases for RAG
chromadb>=0.4.0
pinecone-client>=2.2.0

# Environment and configuration
python-dotenv>=1.0.0

# Code analysis (for Project 4)
pylint>=3.0.0
radon>=6.0.0
ast-tools>=0.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
pydantic>=2.0.0
httpx>=0.25.0
aiohttp>=3.9.0
python-multipart>=0.0.6

# Visualization
matplotlib>=3.8.0
plotly>=5.17.0

# Date/time handling
python-dateutil>=2.8.0

# Google AI SDK (for any additional Gemini features)
google-generativeai>=0.3.0

